add rendering of json resume themes with stimulusReflex
To check to make sure that actioncable was working I created a test
channel. When that was working I decided to keep it so that I can
use it with CableReady to test any straight CableReady actions without
stimulus reflex.

I decided against using a controller for the resume reflex action
due to the fact that I did not use an object with dom_id attached.
It is easier at this point to have the unique dom id 'resume' to find
the element that we are targeting for morph to work on.

I chose morph because it sends over the wire only the html relevent
to the id you give it. If not we would send over the html for the
entire page and increase reaction times.

To conditionally render I used a unique data-id - the url
paramater - and created a method in the view component to return the
json object if triggered by the same id.